requirements.txt
torch==2.5.0
torchvision==0.20.0
torchaudio==2.5.0

numpy
scipy
pandas

opencv-python
pillow
imageio

matplotlib
tqdm

scikit-learn

einops

open3d
trimesh
plyfile

kornia

hydra-core
omegaconf

yacs

tensorboard

huggingface-hub
safetensors

xformers

# Web / UI
gradio==5.17.1
fastapi==0.119.0
flask==3.1.2